Fix Verilog codegen for reg
Previously, we emitted if-else sequences for reg updates.  Recent
improvements to FIRRTL resulted in the emission of explicit mux
networks instead.  While equivalent, doing so reduces QoR, presumably
because ECAD tools are tuned to the habits of manual Verilog coders.

This seems to be a result of WRefs appearing between the regs and muxes.
Chasing down the sources of the WRefs corrects the code generation.

This patch reduces the Rocket pipeline area by about 2% and improves
rocket-chip's Verilator performance by about 8%.